Update python to Python-3.2.6

+++ python/Lib/multiprocessing/util.py
@@ -32,13 +32,11 @@
 # SUCH DAMAGE.
 #
 
-import os
 import itertools
 import weakref
 import atexit
 import threading        # we want threading to install it's
                         # cleanup function before multiprocessing does
-from subprocess import _args_from_interpreter_flags
 
 from multiprocessing.process import current_process, active_children
 
@@ -155,11 +153,11 @@ def _run_after_forkers():
     for (index, ident, func), obj in items:
         try:
             func(obj)
-        except Exception, e:
+        except Exception as e:
             info('after forker raised exception %s', e)
 
 def register_after_fork(obj, func):
-    _afterfork_registry[(_afterfork_counter.next(), id(obj), func)] = obj
+    _afterfork_registry[(next(_afterfork_counter), id(obj), func)] = obj
 
 #
 # Finalization using weakrefs
@@ -174,7 +172,7 @@ class Finalize(object):
     Class which supports object finalization using weakrefs
     '''
     def __init__(self, obj, callback, args=(), kwargs=None, exitpriority=None):
-        assert exitpriority is None or type(exitpriority) in (int, long)
+        assert exitpriority is None or type(exitpriority) is int
 
         if obj is not None:
             self._weakref = weakref.ref(obj, self)
@@ -184,8 +182,7 @@ class Finalize(object):
         self._callback = callback
         self._args = args
         self._kwargs = kwargs or {}
-        self._key = (exitpriority, _finalizer_counter.next())
-        self._pid = os.getpid()
+        self._key = (exitpriority, next(_finalizer_counter))
 
         _finalizer_registry[self._key] = self
 
@@ -198,13 +195,9 @@ class Finalize(object):
         except KeyError:
             sub_debug('finalizer no longer registered')
         else:
-            if self._pid != os.getpid():
-                sub_debug('finalizer ignored because different process')
-                res = None
-            else:
-                sub_debug('finalizer calling %s with args %s and kwargs %s',
-                          self._callback, self._args, self._kwargs)
-                res = self._callback(*self._args, **self._kwargs)
+            sub_debug('finalizer calling %s with args %s and kwargs %s',
+                     self._callback, self._args, self._kwargs)
+            res = self._callback(*self._args, **self._kwargs)
             self._weakref = self._callback = self._args = \
                             self._kwargs = self._key = None
             return res
@@ -265,10 +258,7 @@ def _run_finalizers(minpriority=None):
     else:
         f = lambda p : p[0][0] is not None and p[0][0] >= minpriority
 
-    # Careful: _finalizer_registry may be mutated while this function
-    # is running (either by a GC run or by another thread).
-
-    items = [x for x in _finalizer_registry.items() if f(x)]
+    items = [x for x in list(_finalizer_registry.items()) if f(x)]
     items.sort(reverse=True)
 
     for key, finalizer in items:
@@ -297,38 +287,39 @@ _exiting = False
 def _exit_function(info=info, debug=debug, _run_finalizers=_run_finalizers,
                    active_children=active_children,
                    current_process=current_process):
-    # NB: we hold on to references to functions in the arglist due to the
+    # We hold on to references to functions in the arglist due to the
     # situation described below, where this function is called after this
     # module's globals are destroyed.
 
     global _exiting
 
-    info('process shutting down')
-    debug('running all "atexit" finalizers with priority >= 0')
-    _run_finalizers(0)
+    if not _exiting:
+        _exiting = True
 
-    if current_process() is not None:
-        # NB: we check if the current process is None here because if
-        # it's None, any call to ``active_children()`` will throw an
-        # AttributeError (active_children winds up trying to get
-        # attributes from util._current_process).  This happens in a
-        # variety of shutdown circumstances that are not well-understood
-        # because module-scope variables are not apparently supposed to
-        # be destroyed until after this function is called.  However,
-        # they are indeed destroyed before this function is called.  See
-        # issues 9775 and 15881.  Also related: 4106, 9205, and 9207.
+        info('process shutting down')
+        debug('running all "atexit" finalizers with priority >= 0')
+        _run_finalizers(0)
+        if current_process() is not None:
+            # We check if the current process is None here because if
+            # it's None, any call to ``active_children()`` will raise an
+            # AttributeError (active_children winds up trying to get
+            # attributes from util._current_process).  This happens in a
+            # variety of shutdown circumstances that are not well-understood
+            # because module-scope variables are not apparently supposed to
+            # be destroyed until after this function is called.  However,
+            # they are indeed destroyed before this function is called.  See
+            # issues #9775 and #15881.  Also related: #4106, #9205, and #9207.
+            for p in active_children():
+                if p._daemonic:
+                    info('calling terminate() for daemon %s', p.name)
+                    p._popen.terminate()
 
-        for p in active_children():
-            if p._daemonic:
-                info('calling terminate() for daemon %s', p.name)
-                p._popen.terminate()
+            for p in active_children():
+                info('calling join() for process %s', p.name)
+                p.join()
 
-        for p in active_children():
-            info('calling join() for process %s', p.name)
-            p.join()
-
-    debug('running the remaining "atexit" finalizers')
-    _run_finalizers()
+        debug('running the remaining "atexit" finalizers')
+        _run_finalizers()
 
 atexit.register(_exit_function)
 
@@ -338,13 +329,10 @@ atexit.register(_exit_function)
 
 class ForkAwareThreadLock(object):
     def __init__(self):
-        self._reset()
-        register_after_fork(self, ForkAwareThreadLock._reset)
-
-    def _reset(self):
         self._lock = threading.Lock()
         self.acquire = self._lock.acquire
         self.release = self._lock.release
+        register_after_fork(self, ForkAwareThreadLock.__init__)
 
 class ForkAwareLocal(threading.local):
     def __init__(self):
